part1 complete but untested

+++ libft/ft_memchr.c
@@ -6,18 +6,18 @@
 /*   By: pehenriq <pehenriq@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/04 16:34:31 by pehenriq          #+#    #+#             */
-/*   Updated: 2021/06/04 16:58:16 by pehenriq         ###   ########.fr       */
+/*   Updated: 2021/06/06 11:56:55 by pehenriq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (int) n && (unsigned char *)s[i] != (unsigned char) c)
+	while (i < n && (unsigned char *)s[i] != (unsigned char) c)
 		i++;
-	if (i == (int) n)
+	if (i == n)
 		return (0);
 	return ((unsigned char *)&s[i]);
 }
